Fix ADI dataset name
